Enable multi platform builds (#45)
Currently the GHGA services are only built for x86_64/amd64 architectures. Not only does this limit the adoptability on more cost-efficient arm nodes, but also poses a problem for developers operating on arm cpus (such as newer MacBooks).

This can be fixed by adding the option platforms: linux/amd64,linux/arm64 to the docer/build-push-action.